requirements.txt
Flask==2.3.2
Flask-HTTPAuth==4.8.0
appengine-python-standard>=1.0.0
google-cloud-secret-manager==2.16.1